refactor(runs): Reorganize file structure
